Description: js_textfield告警清理 IssueNo: https://gitee.com/openharmony/arkui_ace_engine/issues/IBLZA8 Feature or Bugfix: Feature Binary Source:No Signed-off-by: chenjiafeng <chenjiafeng10@huawei.com>

+++ frameworks/bridge/declarative_frontend/jsview/js_text.cpp
@@ -949,7 +949,7 @@ void JSText::BindSelectionMenu(const JSCallbackInfo& info)
     // SelectionMenuOptions
     NG::SelectMenuParam menuParam;
     menuParam.isValid = isValidTextSpanType;
-    if (info.Length() > resquiredParameterCount) {
+    if (info.Length() > static_cast<uint32_t>(resquiredParameterCount)) {
         JSRef<JSVal> argsMenuOptions = info[resquiredParameterCount];
         if (argsMenuOptions->IsObject()) {
             ParseMenuParam(info, argsMenuOptions, menuParam);
